refactor(node): Move `Process` definition to `process.d.ts`
