+++ types/node/ts3.2/base.d.ts
@@ -19,5 +19,5 @@
 // TypeScript 3.2-specific augmentations:
 /// <reference path="buffer.d.ts" />
 /// <reference path="fs.d.ts" />
+/// <reference path="process.d.ts" />
 /// <reference path="util.d.ts" />
-/// <reference path="globals.d.ts" />
