--- types/node/ts3.2/globals.d.ts
@@ -1,8 +1,0 @@
-// tslint:disable-next-line:no-bad-reference
-/// <reference path="../globals.d.ts" />
-
-declare namespace NodeJS {
-    interface HRTime {
-        bigint(): bigint;
-    }
-}
