+++ types/node/ts3.2/process.d.ts
@@ -1,0 +1,12 @@
+// tslint:disable-next-line:no-bad-reference
+/// <reference path="../process.d.ts" />
+
+declare module 'process' {
+    global {
+        namespace NodeJS {
+            interface HRTime {
+                bigint(): bigint;
+            }
+        }
+    }
+}
